Move ESP32 OTA updates app notifications into main.c
